Fix some barcodes not detected on newer iphones
The method AVCaptureDevice.DevicesWithMediaType is deprecated and did not returned triple and dual camera devices, making it harder to scan barcodes due to the lower quality of the camera device returned. Switching to AVCaptureDeviceDiscoverySession allows to ask for a list of ordered device types that our application would like to use.